{"schema":"decodex/commit/1","summary":"Extract native scroll-capture wheel input","authority":"manual"}
