+++ docs/reference/host-core-reset.md
@@ -70,8 +70,10 @@ The current native-host Swift split is:
 - `CaptureSessionController.swift` plus `CaptureSessionController+*.swift`: the Swift host-session
   coordinator around the Rust `RsnapHostSession`. The base file owns shared controller state and
   lifecycle hooks; the extensions split live capture/input, frozen selection interactions,
-  host-request draining, native scroll-capture sampling, copy/save/export effects, Vision OCR, and
+  host-request draining, native scroll-capture lifecycle, copy/save/export effects, Vision OCR, and
   runtime teardown/window helpers.
+- `NativeScrollCaptureWheelInput.swift`: native scroll-capture wheel interception, global monitor
+  lifecycle, forwarded CGEvent posting, queued forwarded-delta draining, and motion-hint updates.
 - `NativeScrollCaptureObservationPipeline.swift`: native scroll-capture sample batching, fallback
   sample adaptation, Rust scroll-observation calls, and preview export refresh packaging. It keeps
   ordered frame acquisition and AppKit scheduling in Swift while leaving stitching decisions in
